ambiente-testes/backend-python/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
pydantic[email]==2.6.3
pymongo==4.6.2
motor==3.3.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
pytest==8.0.0
httpx==0.27.0
pytest-cov==4.1.0